+++ deerlab/classes.py
@@ -8,197 +8,6 @@ from scipy.signal import fftconvolve
 from scipy.linalg import block_diag
 from scipy.optimize import brentq
 from scipy.interpolate import interp1d
-import difflib
-import inspect 
-
-
-class FitResult(dict):
-# ========================================================================
-    r""" Represents the results of either the :ref:`fit` or :ref:`snlls` functions. 
-    Depending on which function is used, the attributes and methods of the FitResult will be different.
-
- 
-    Attributes
-    ----------
-    model : ndarray
-        The fitted model response.
-    modelUncert : 
-        Uncertainty quantification of the fitted model response.
-    regparam : float scalar
-        Regularization parameter used in the fit.
-    noiselvl: ndarray
-        Estimated noise level of the data or user-provided noise level.
-    success : bool
-        Whether or not the optimizer exited successfully.
-    cost : float
-        Value of the cost function at the solution.
-    residuals : ndarray
-        Vector of residuals at the solution.
-    stats : dict
-        Goodness of fit statistical estimators:
-
-        * ``stats['chi2red']`` - Reduced \chi^2 test
-        * ``stats['r2']`` - R^2 test
-        * ``stats['rmsd']`` - Root-mean squared deviation (RMSD)
-        * ``stats['aic']`` - Akaike information criterion
-        * ``stats['aicc']`` - Corrected Akaike information criterion
-        * ``stats['bic']`` - Bayesian information criterion
-
-    Methods
-    -------
-    plot(axis=None,xlabel='',gof=False)
-        Function to display the results. It will display the fitted data. A 
-        vector for the x-axis and its label can be specified by calling 
-        FitResult.plot(axis=x,xlabel='xlabel'). A set of goodness-of-fit plots 
-        can be displayed by enabling the gof option by calling FitResult.plot(gof=True).
-    
-    """
-
-    def __getattr__(self, attr):
-        try:
-            return self[attr]
-        except KeyError:
-            errstr = f"The results object has no attribute '{attr}'."
-            attributes = [key for key in self.keys()]
-            proposal = difflib.get_close_matches(attr, attributes)
-            if len(proposal)>0:
-                errstr += f' \n Did you mean: {proposal} ?'
-            raise AttributeError(errstr)
-
-    __setattr__ = dict.__setitem__
-    __delattr__ = dict.__delitem__
-
-    def __str__(self): 
-        return self._summary
-
-    def __repr__(self):
-        if self.keys():
-            m = max(map(len, list(self.keys()))) + 1
-            return '\n'.join([k.rjust(m) + ': ' + repr(v)
-                              for k, v in sorted(self.items())])
-        else:
-            return self.__class__.__name__ + "()"
-
-    def __dir__(self):
-        return list(self.keys())
-    
-
-    def _extarct_params_from_model(self, model):
-        if callable(model):
-            try:
-                modelparam = model._parameter_list('vector')
-            except AttributeError:
-                modelparam = inspect.getfullargspec(model).args
-
-        if not hasattr(self,'param'):
-            raise ValueError('The fit object does not contain any fitted parameters.')
-
-        # Enforce model normalization
-        normalization = False
-        normfactor_keys = []
-        for key in model._parameter_list():
-            param = getattr(model,key)
-            if np.all(param.linear):
-                if param.normalization is not None:
-                    normfactor_key = f'{key}_scale'
-                    normfactor_keys.append(normfactor_key)
-                    model.addnonlinear(normfactor_key,lb=-np.inf,ub=np.inf,par0=1,description=f'Normalization factor of {key}')
-                    getattr(model,normfactor_key).freeze(1)
-                    normalization = True
-                    
-
-        # Get some basic information on the parameter vector
-        modelparam = model._parameter_list(order='vector')
-        param_idx = [[] for _ in model._parameter_list('vector')]
-        idxprev = 0
-        for islinear in [False,True]:
-            for n,param in enumerate(model._parameter_list('vector')):
-                if np.all(getattr(model,param).linear == islinear):
-                    N = len(np.atleast_1d(getattr(model,param).idx))
-                    param_idx[n] = np.arange(idxprev,idxprev + N)
-                    idxprev += N  
-
-        params = {key : fitvalue if len(fitvalue)>1 else fitvalue[0] for key,fitvalue in zip(modelparam,[self.param[idx] for idx in param_idx])}
-        # Check that all parameters are in the fit object
-        for param in modelparam:
-            if not param in params: 
-                raise KeyError(f'The fit object does not contain the {param} parameter.')
-
-        return modelparam, params, param_idx
-
-    def evaluate(self, model, *constants):
-    # ----------------------------------------------------------------------------
-        """
-        Evaluate a model at the fitted parameter values. 
-
-        Takes a model object or callable function model to be evaluated. All the 
-        parameters in the model or in the callable definition must match their 
-        corresponding parameter names in the FitResult object. Any model 
-        constants present required by the model must be specified as a second 
-        argument constants. It returns the model's response at the fitted 
-        parameter values as an ndarray.
-
-        Parameters
-        ----------
-
-        model : :ref:`Model` or callable
-            Model object or callable function to be evaluated. All the parameters in the model or in the callable definition
-            must match their corresponding parameter names in the ``FitResult`` object.   
-        constants : array_like 
-            Any model constants present required by the model.  
-        
-        Returns
-        -------
-
-        response : array_like 
-            Model response at the fitted parameter values. 
-        """
-
-        modelparam, params, _ = self._extarct_params_from_model(model)
-        parameters = {param: params[param] for param in modelparam}
-
-        # Evaluate the input model
-        response = model(*constants,**parameters)
-        return response
-    
-    def propagate(self, model, *constants, lb=None, ub=None):
-        """
-        Propagate the uncertainty in the fit results to a model's response.
-
-        Takes a model object or callable function model to be 
-        evaluated. All the parameters in the model or in the callable definition 
-        must match their corresponding parameter names in the FitResult object.
-        Any model constants present required by the model must be specified as 
-        a second argument constants. The lower bounds lb and upper bounds ub of
-        the model's response can be specified as a third and fourth argument 
-        respectively. It returns the model's response uncertainty 
-        quantification as a UQResult object.
-        Parameters
-        ----------
-
-        model : :ref:`Model` or callable
-            Model object or callable function to be evaluated. All the parameters in the model or in the callable definition
-            must match their corresponding parameter names in the ``FitResult`` object.   
-        constants : array_like 
-            Model constants. 
-        lb : array_like, optional 
-            Lower bounds of the model response.
-        ub : array_like, optional 
-            Upper bounds of the model response.   
-
-        Returns
-        -------
-
-        responseUncert : :ref:`UQResult`
-            Uncertainty quantification of the model's response.
-        """
-
-        modelparam,_, param_idx = self._extarct_params_from_model(model)
-        # Determine the indices of the subset of parameters the model depends on
-        subset = [param_idx[np.where(np.asarray(modelparam)==param)[0][0]] for param in modelparam]
-        # Propagate the uncertainty from that subset to the model
-        modeluq = self.paramUncert.propagate(lambda param: model(*constants,*[param[s] for s in subset]),lb,ub)
-        return modeluq
 
 
 # =========================================================================
